Added Dispatcher. Removed SQLite DB.
Added Gateway and Fake Gateway implementation.
Added Manager and implementation.
Added Application and Dependency Manager.
Tested manager.

+++ app/src/main/java/com/challenge/lanchonete/models/Ingredient.kt
@@ -3,5 +3,5 @@ package com.challenge.lanchonete.models
 data class Ingredient(
     val id: Int,
     val name: String = "",
-    val value: Int = 0
+    val value: Double = 0.0
 )
